# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation about projekt SasiadWPotrzebie",
        "description": "sasiadwpotrzebie api",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9000",
            "description": "local server"
        },
        {
            "url": "http://localhost:9000",
            "description": "staging server"
        },
        {
            "url": "http://localhost:9000",
            "description": "production server"
        }
    ],
    "paths": {
...
        "/api/product-categories": {
            "get": {
                "tags": [
                    "Product Categories"
                ],
                "summary": "Retrieve all product categories",
                "description": "Get all product categories",
                "operationId": "dc793ac98441da2fcda6b5d04eb78104",
                "responses": {
                    "200": {
                        "description": "List of all product categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Electronics"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-09-22T10:20:30Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-09-23T15:30:45Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/product-categories/{id}": {
            "get": {
                "tags": [
                    "Product Categories"
                ],
                "summary": "Retrieve a product category by its ID",
                "description": "Get product category by id",
                "operationId": "9a78d43b83a21694ba25959e9f34d872",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product category details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Clothing"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-09-25T11:15:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-09-26T17:45:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product category not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/product-categories/create": {
            "post": {
                "tags": [
                    "Product Categories"
                ],
                "summary": "Create a new product category",
                "description": "Create product category",
                "operationId": "0db2c2c902778ea7b21b381067147b86",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Books"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product category created successfully"
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/product-categories/{id}/update": {
            "patch": {
                "tags": [
                    "Product Categories"
                ],
                "summary": "Update a product category by its ID",
                "description": "Update product category by id",
                "operationId": "ac6adfb5c74c91c32d6f8e298dfbcb73",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Category"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product category updated successfully"
                    },
                    "404": {
                        "description": "Product category not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/product-categories/{id}/delete": {
            "delete": {
                "tags": [
                    "Product Categories"
                ],
                "summary": "Delete a product category by its ID",
                "description": "Delete product category by id",
                "operationId": "a725d8122b982fa04aa569b99ac76e21",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product category deleted successfully"
                    },
                    "404": {
                        "description": "Product category not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}